move the running server logic to modkit::bootstrap
