+++ apps/hyperspot-server/src/main.rs
@@ -2,26 +2,16 @@ mod registered_modules;
 
 use anyhow::Result;
 use clap::{Parser, Subcommand};
-use figment::Figment;
 use mimalloc::MiMalloc;
-use modkit::LocalProcessBackend;
-use modkit::bootstrap::config::{
-    AppConfig, RuntimeKind, dump_effective_modules_config_json, dump_effective_modules_config_yaml,
-    get_module_runtime_config, list_module_names, render_module_config_for_oop,
+use modkit::bootstrap::{
+    AppConfig, dump_effective_modules_config_json, dump_effective_modules_config_yaml,
+    host::init_logging_unified, list_module_names, run_server,
 };
-use modkit::bootstrap::host::{init_logging_unified, normalize_path};
-use tokio_util::sync::CancellationToken;
 
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
+use std::path::PathBuf;
 
 #[global_allocator]
 static GLOBAL: MiMalloc = MiMalloc;
-
-// Bring runner types & our per-module DB factory
-use modkit::runtime::{
-    DbOptions, OopModuleSpawnConfig, OopSpawnOptions, RunOptions, ShutdownOptions, run, shutdown,
-};
 
 /// `HyperSpot` Server - modular platform for AI services
 #[derive(Parser)]
@@ -105,11 +95,7 @@ async fn main() -> Result<()> {
 
     // Initialize logging + otel in one Registry
     let logging_config = config.logging.clone().unwrap_or_default();
-    init_logging_unified(
-        &logging_config,
-        Path::new(&config.server.home_dir),
-        otel_layer,
-    );
+    init_logging_unified(&logging_config, &config.server.home_dir, otel_layer);
 
     // One-time connectivity probe
     #[cfg(feature = "otel")]
@@ -163,7 +149,7 @@ async fn main() -> Result<()> {
 
     // Dispatch subcommands (default: run)
     match cli.command.as_ref().unwrap_or(&Commands::Run) {
-        Commands::Run => run_server(config, cli).await,
+        Commands::Run => run_server(config).await,
         Commands::Check => check_config(&config),
     }
 }
@@ -174,25 +160,6 @@ fn check_config(config: &AppConfig) -> Result<()> {
     println!("Configuration is valid");
     println!("{}", config.to_yaml()?);
     Ok(())
-}
-
-/// Create a Figment from the loaded `AppConfig` for use with `DbManager`.
-fn create_figment_from_config(config: &AppConfig) -> Figment {
-    use figment::providers::Serialized;
-
-    // Convert the AppConfig back to a Figment that DbManager can use
-    // We serialize the config and then parse it back as a Figment
-    Figment::new().merge(Serialized::defaults(config))
-}
-
-/// Create a mock Figment for testing with in-memory `SQLite` databases.
-fn create_mock_figment(config: &AppConfig) -> Figment {
-    use figment::providers::Serialized;
-
-    let mut mock_config = config.clone();
-    override_modules_with_mock_db(&mut mock_config);
-
-    Figment::new().merge(Serialized::defaults(mock_config))
 }
 
 /// Override all module database configurations with in-memory `SQLite`
@@ -214,168 +181,4 @@ fn override_modules_with_mock_db(config: &mut AppConfig) {
             );
         }
     }
-}
-
-/// Resolve database options based on configuration and args
-fn resolve_db_options(config: &AppConfig, args: &Cli) -> Result<DbOptions> {
-    if config.database.is_none() {
-        tracing::warn!("No global database section found; running without databases");
-        return Ok(DbOptions::None);
-    }
-
-    if args.mock {
-        return create_mock_db_manager(config);
-    }
-
-    create_standard_db_manager(config)
-}
-
-/// Create mock `DbManager` with in-memory `SQLite`
-fn create_mock_db_manager(config: &AppConfig) -> Result<DbOptions> {
-    tracing::info!("Mock mode enabled: using in-memory SQLite for all modules");
-    let mock_figment = create_mock_figment(config);
-    let home_dir = PathBuf::from(&config.server.home_dir);
-    let db_manager = Arc::new(modkit_db::DbManager::from_figment(mock_figment, home_dir)?);
-    Ok(DbOptions::Manager(db_manager))
-}
-
-/// Create standard `DbManager` from configuration
-fn create_standard_db_manager(config: &AppConfig) -> Result<DbOptions> {
-    tracing::info!("Using DbManager with Figment-based configuration");
-    let figment = create_figment_from_config(config);
-    let home_dir = PathBuf::from(&config.server.home_dir);
-    let db_manager = Arc::new(modkit_db::DbManager::from_figment(figment, home_dir)?);
-    Ok(DbOptions::Manager(db_manager))
-}
-
-async fn run_server(config: AppConfig, args: Cli) -> Result<()> {
-    tracing::info!("Initializing modules...");
-
-    // Generate process-level instance ID once at startup.
-    // This is shared by all modules in this process.
-    let instance_id = uuid::Uuid::new_v4();
-    tracing::info!(instance_id = %instance_id, "Generated process instance ID");
-
-    // Create root cancellation token for the entire process.
-    // This token drives shutdown for the module runtime and all lifecycle/stateful modules.
-    let cancel = CancellationToken::new();
-
-    // Hook OS signals to the root token at the host level.
-    // This replaces the use of ShutdownOptions::Signals inside the runtime.
-    let cancel_for_signals = cancel.clone();
-    tokio::spawn(async move {
-        match shutdown::wait_for_shutdown().await {
-            Ok(()) => {
-                tracing::info!("shutdown: signal received in master host");
-            }
-            Err(e) => {
-                tracing::warn!(
-                    error = %e,
-                    "shutdown: primary waiter failed in master host, falling back to ctrl_c()"
-                );
-                let _ = tokio::signal::ctrl_c().await;
-            }
-        }
-        cancel_for_signals.cancel();
-    });
-
-    // Build config provider and resolve database options
-    let db_options = resolve_db_options(&config, &args)?;
-
-    // Create OoP backend with cancellation token - it will auto-shutdown all processes on cancel
-    let oop_backend = LocalProcessBackend::new(cancel.clone());
-
-    // Build OoP spawn configuration
-    let oop_options = build_oop_spawn_options(&config, &args, oop_backend)?;
-
-    // Run the ModKit runtime with the root cancellation token.
-    // Shutdown is driven by the signal handler spawned above, not by ShutdownOptions::Signals.
-    // OoP modules are spawned after the start phase (once grpc_hub has bound its port).
-    let run_options = RunOptions {
-        modules_cfg: Arc::new(config),
-        db: db_options,
-        shutdown: ShutdownOptions::Token(cancel.clone()),
-        clients: vec![],
-        instance_id,
-        oop: oop_options,
-    };
-
-    let result = run(run_options).await;
-
-    // Graceful shutdown - flush any remaining traces
-    #[cfg(feature = "otel")]
-    modkit::telemetry::init::shutdown_tracing();
-
-    result
-}
-
-/// Build `OoP` spawn configuration from `AppConfig`.
-///
-/// This collects all modules with `type=oop` and prepares their spawn configuration.
-/// The actual spawning happens in the `HostRuntime` after the start phase.
-fn build_oop_spawn_options(
-    config: &AppConfig,
-    _args: &Cli,
-    backend: LocalProcessBackend,
-) -> Result<Option<OopSpawnOptions>> {
-    let home_dir = PathBuf::from(&config.server.home_dir);
-    let mut modules = Vec::new();
-
-    for module_name in config.modules.keys() {
-        if let Some(spawn_config) = try_build_oop_module_config(config, module_name, &home_dir)? {
-            modules.push(spawn_config);
-        }
-    }
-
-    if modules.is_empty() {
-        Ok(None)
-    } else {
-        tracing::info!(count = modules.len(), "Prepared OoP modules for spawning");
-        Ok(Some(OopSpawnOptions {
-            modules,
-            backend: Box::new(backend),
-        }))
-    }
-}
-
-/// Try to build `OoP` module spawn config if module is of type `OoP`
-fn try_build_oop_module_config(
-    config: &AppConfig,
-    module_name: &str,
-    home_dir: &Path,
-) -> Result<Option<OopModuleSpawnConfig>> {
-    let Some(runtime_cfg) = get_module_runtime_config(config, module_name)? else {
-        return Ok(None);
-    };
-
-    if runtime_cfg.mod_type != RuntimeKind::Oop {
-        return Ok(None);
-    }
-
-    let exec_cfg = runtime_cfg.execution.as_ref().ok_or_else(|| {
-        anyhow::anyhow!("module '{module_name}' is type=oop but execution config is missing")
-    })?;
-
-    let binary = normalize_path(&exec_cfg.executable_path)?;
-    let spawn_args = exec_cfg.args.clone();
-    let env = exec_cfg.environment.clone();
-
-    // Render the complete module config (with resolved DB)
-    let rendered_config = render_module_config_for_oop(config, module_name, home_dir)?;
-    let rendered_json = rendered_config.to_json()?;
-
-    tracing::debug!(
-        module = %module_name,
-        "Prepared OoP module config: db={}",
-        rendered_config.database.is_some()
-    );
-
-    Ok(Some(OopModuleSpawnConfig {
-        module_name: module_name.to_owned(),
-        binary,
-        args: spawn_args,
-        env,
-        working_directory: exec_cfg.working_directory.clone(),
-        rendered_config_json: rendered_json,
-    }))
 }
