+++ libs/modkit/src/bootstrap/mod.rs
@@ -28,3 +28,6 @@ pub use config::{
 
 // Re-export host types for convenience
 pub use oop::{OopRunOptions, run_oop_with_options};
+
+mod run;
+pub use run::run_server;
